chore(plugins): take comments
