+++ src/kili/cli/project/__init__.py
@@ -10,7 +10,6 @@ from kili.cli.project.import_ import import_assets
 from kili.cli.project.label import import_labels
 from kili.cli.project.list_ import list_projects
 from kili.cli.project.member import member
-from kili.cli.project.plugin import upload_plugin
 
 
 @click.group(context_settings=CONTEXT_SETTINGS)
@@ -25,4 +24,3 @@ project.add_command(import_labels, name="label")
 project.add_command(list_projects, name="list")
 project.add_command(member, name="member")
 project.add_command(export_labels, name="export")
-project.add_command(upload_plugin, name="upload_plugin")
