+++ src/kili/graphql_client.py
@@ -90,7 +90,7 @@ class GraphQLClient:
             req = None
             try:
                 try:
-                    number_of_trials = int(os.getenv("KILI_SDK_TRIALS_NUMBER", "1"))
+                    number_of_trials = int(os.getenv("KILI_SDK_TRIALS_NUMBER", "10"))
                 except ValueError:
                     number_of_trials = 1
                 for trial_number in range(number_of_trials):
