+++ src/kili/helpers.py
@@ -10,14 +10,13 @@ import os
 import re
 import warnings
 from json import dumps, loads
-from pathlib import Path
 from typing import Callable, List, Optional, Type, TypeVar
 
 import pyparsing as pp
 import requests
 from typing_extensions import TypedDict, get_args, get_origin, is_typeddict
 
-from kili.constants import mime_extensions_for_IV2, mime_extensions_for_py_scripts
+from kili.constants import mime_extensions_for_IV2
 from kili.exceptions import GraphQLError, NonExistingFieldError
 
 T = TypeVar("T")
@@ -398,25 +397,5 @@ def check_file_mime_type(path: str, input_type: str, raise_error=True) -> bool:
             f"File mime type for {path} is {mime_type} and does not correspond "
             "to the type of the project. "
             f"File mime type should be one of {mime_extensions_for_IV2[input_type]}"
-        )
-    return correct_mime_type
-
-
-def check_file_is_py(path: Path, verbose: bool = True) -> bool:
-    """
-    Returns true if the mime type of the file corresponds to a python file
-    """
-    mime_type = get_data_type(path.as_posix())
-
-    if not (mime_extensions_for_py_scripts and mime_type):
-        return False
-
-    correct_mime_type = mime_type in mime_extensions_for_py_scripts
-
-    if verbose and not correct_mime_type:
-        print(
-            f"File mime type for {path} is {mime_type} and does not correspond"
-            "to the type of the project. "
-            f"File mime type should be one of {mime_extensions_for_py_scripts}"
         )
     return correct_mime_type
