--- src/kili/mutations/plugins/queries.py
@@ -1,31 +1,0 @@
-"""
-Queries of plugins mutations
-"""
-
-GQL_ACTIVATE_PLUGIN_ON_PROJECT = """
-mutation(
-  $pluginName: String!
-  $projectId: ID!
-  ) {
-  data: activatePluginOnProject(
-    where: {
-      pluginName: $pluginName
-      projectId: $projectId
-    }
-  )
-}
-"""
-
-GQL_DEACTIVATE_PLUGIN_ON_PROJECT = """
-mutation(
-  $pluginName: String!
-  $projectId: ID!
-  ) {
-  data: deactivatePluginOnProject(
-    where: {
-      pluginName: $pluginName
-      projectId: $projectId
-    }
-  )
-}
-"""
